% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_interval_calculator.R
\name{conf_interval_calculator}
\alias{conf_interval_calculator}
\title{conf_interval_calculator}
\usage{
conf_interval_calculator(
  probability = svm_confidence_table$SVM_PROB,
  prediction = svm_confidence_table$SVM_LABEL,
  confidence_interval = 0.95
)
}
\description{
conf_interval_calculator
}
