% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_confidence_table.R
\name{svm_confidence_table}
\alias{svm_confidence_table}
\title{svm_confidence_table}
\usage{
svm_confidence_table(
  x,
  actual_label,
  model,
  original_matrix = dtMatrix,
  have_labels = T
)
}
\description{
svm_confidence_table
}
