% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor_dataframe_generator.R
\name{vendor_dataframe_generator}
\alias{vendor_dataframe_generator}
\title{vendor_dataframe_generator}
\usage{
vendor_dataframe_generator(labels, vendor)
}
\description{
vendor_dataframe_generator
}
